Rename all userBases to user
