+++ web/.gitignore
@@ -23,6 +23,7 @@ yarn-debug.log*
 yarn-error.log*
 
 #amplify
+amplify/
 amplify/\#current-cloud-backend
 amplify/.config/local-*
 amplify/mock-data
