+++ web/src/actions/basesActions.js
@@ -1,0 +1,8 @@
+import { BASE_SELECTED } from './types';
+
+export const selectBase = (base) => {
+  return {
+    type: BASE_SELECTED,
+    payload: base,
+  };
+};
