+++ web/src/actions/types.js
@@ -1,2 +1,2 @@
 export const BASE_SELECTED = 'BASE_SELECTED';
-export const FETCH_USER_BASES = 'FETCH_USER_BASES';
+export const FETCH_USER = 'FETCH_USER';
